feat(past-sessions): showing past session and profile to user that is logged in

+++ src/app/dashboard/past-sessions/reports/key-stats/participants/participants.component.ts
@@ -1,5 +1,8 @@
 import { Component, Input, OnChanges, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material';
+import { ContextService } from 'src/app/services';
 import { PastSessionsService } from 'src/app/services/past-sessions.service';
+import { ErrorMessageDialogComponent } from 'src/app/shared';
 
 @Component({
   selector: 'benji-participants',
@@ -8,13 +11,24 @@ import { PastSessionsService } from 'src/app/services/past-sessions.service';
 })
 export class ParticipantsComponent implements OnInit, OnChanges {
   @Input() data: any = { joined_users: [] };
+  allowAllSelection;
   selected = [];
   participants = [];
-  constructor(private pastSessionService: PastSessionsService) {}
+  dialogRef;
+  constructor(
+    private pastSessionService: PastSessionsService,
+    private contextService: ContextService,
+    private dialog: MatDialog
+  ) {}
 
   ngOnInit() {
     this.pastSessionService.filteredInUsers$.subscribe(list => {
       this.selected = list;
+    });
+    this.contextService.user$.subscribe(user => {
+      if (user.local_admin_permission) {
+        this.allowAllSelection = true;
+      }
     });
   }
 
@@ -29,10 +43,52 @@ export class ParticipantsComponent implements OnInit, OnChanges {
   }
 
   participantClicked(event, participant) {
+    // if (this.allowAllSelection) {
     this.pastSessionService.addToFilteredInList(participant.id);
+    // } else {
+    //   const msg =
+    //     'You do not have permission to view these reports' +
+    //     '\nContact your manager';
+    //   this.dialogRef = this.dialog
+    //     .open(ErrorMessageDialogComponent, {
+    //       data: {
+    //         confirmationMessage: msg
+    //       },
+    //       disableClose: true,
+    //       panelClass: 'dashboard-dialog'
+    //     })
+    //     .afterClosed()
+    //     .subscribe(res => {
+    //       if (res) {
+    //         console.log('br');
+    //       }
+    //     });
+    //   // you don't have admin permissions
+    //   // contact your manager or benji
+    // }
   }
 
   showOnly(participant) {
-    this.pastSessionService.removeAllBut(participant.id);
+    if (this.allowAllSelection) {
+      this.pastSessionService.removeAllBut(participant.id);
+    } else {
+      const msg =
+        'You do not have permission to view these reports' +
+        '\nContact your manager.';
+      this.dialogRef = this.dialog
+        .open(ErrorMessageDialogComponent, {
+          data: {
+            errorMessage: msg
+          },
+          disableClose: true,
+          panelClass: 'dashboard-dialog'
+        })
+        .afterClosed()
+        .subscribe(res => {
+          if (res) {
+            console.log('br');
+          }
+        });
+    }
   }
 }
